requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
torch==2.4.1
torchvision==0.19.1
transformers>=4.57.1
accelerate>=0.33.0
peft>=0.17.1
huggingface-hub>=0.23.0
pydantic==2.5.0
opencv-python-headless==4.8.1.78
numpy>=1.26.4,<2.0.0
python-dotenv==1.0.0
qwen-vl-utils>=0.0.6
